get course by courseId claim
