+++ LMS.API/Features/Courses/GetCourse/Mapper.cs
@@ -1,0 +1,15 @@
+﻿namespace Courses.GetCourse;
+
+public class Mapper : ResponseMapper<CourseResponseModel, Course>
+{
+    public override CourseResponseModel FromEntity(Course c)
+    {
+        return new CourseResponseModel()
+        {
+            Name = c.Name,
+            Description = c.Name,
+            StartDate = c.StartDate,
+            EndDate = c.EndDate,
+        };
+    }
+}
